+++ CHANGELOG.md
@@ -14,6 +14,7 @@
         * Integer fields to store an integer in the given range.
         * Time fields which can store timestamps.
     * Removed `client.ImportValueField` function. `client.ImportField` function imports both set and integer fields, depending on the type of the field.
+    * Removed index and field validation. The validation is done only on the server side. `schema.Index` and `index.Field` functions do not return `error` values.
   
 * **v0.9.0** (2018-05-10)
     * Compatible with Pilosa 0.9.